Changed the data types of the MortgageAffordabilityRequest to BigDecimal to avoid precision loss

+++ src/main/java/nl/ing/mortgage/affordability/mortgage/check/MortgageAffordabilityRequest.java
@@ -1,9 +1,11 @@
 package nl.ing.mortgage.affordability.mortgage.check;
 
+import java.math.BigDecimal;
+
 public record MortgageAffordabilityRequest(
-        int income,
+        BigDecimal income,
         int maturityPeriod,
-        long loanValue,
-        long homeValue
+        BigDecimal loanValue,
+        BigDecimal homeValue
 ) {
 }
